refactor(controller): split doctor.ts into a barrel over doctor/
1245 lines, of which ~1000 were ten independent check functions and the
review pass. The section roster and its runners stay in doctor.ts (96
lines now); the rest moves out, grouped by what each check inspects:

  types.ts            findings, fix ids, the settings slice checks inspect
  util.ts             the `safe` wrapper, error classification, formatting
  cache.ts            the last-run cache the admin panel hydrates from
  checks-services.ts  LLM / Navidrome / broadcast / TTS
  checks-station.ts   capabilities, host resources, tuning coherence
  checks-content.ts   M3Us, state-dir size, first-run setup
  review.ts           DJ Doc's verdict pass over the assembled report

Grouping by subject rather than one file per check keeps three readable
modules instead of ten stubs. cache.ts had to sort before review.ts —
reviewReport writes its verdict straight into the cache, which was the
only back-edge in the section order.

doctor.ts re-exports everything, so `import * as doctor from './doctor.js'`
in routes/doctor.ts and broadcast/scheduler.ts is unchanged.

No behaviour change. Controller lint clean, all 52 test files pass.
