+++ controller/src/doctor.ts
@@ -3,6 +3,18 @@
 // returns a structured report; reviewReport() asks the configured LLM to turn
 // that report into a prioritized, plain-English read. No rendering here.
 //
+// This module owns the section roster and the runners that drive it; the parts
+// live in ./doctor/ and are re-exported below, so `from './doctor.js'` reaches
+// the whole surface:
+//
+//   types.ts            findings, fix ids, the settings slice checks inspect
+//   util.ts             the `safe` wrapper, error classification, formatting
+//   cache.ts            the last-run cache the admin panel hydrates from
+//   checks-services.ts  LLM / Navidrome / broadcast / TTS
+//   checks-station.ts   capabilities, host resources, tuning coherence
+//   checks-content.ts   M3Us, state-dir size, first-run setup
+//   review.ts           DJ Doc's verdict pass over the assembled report
+//
 // The Finding/DoctorReport shape mirrors cli/src/doctor.ts (same ok|warn|fail|
 // skip vocabulary) so the two surfaces stay legible together — extended with an
 // optional `fix` descriptor that the panel maps to an existing admin endpoint.
@@ -10,117 +22,20 @@
 // probe, Subsonic ping, listener monitor, TTS routing, library stats); nothing
 // here adds new probing infrastructure.
 
-import { readFile, readdir, stat, writeFile } from 'node:fs/promises';
-import type { Dirent } from 'node:fs';
-import { z } from 'zod';
-import { config, STATE_DIR } from './config.js';
 import * as settings from './settings.js';
-import * as subsonic from './music/subsonic.js';
-import * as subsonicLog from './music/subsonic-log.js';
-import * as library from './music/library.js';
-import * as embeddings from './music/embeddings.js';
-import * as tts from './audio/tts.js';
-import { getStreamStatus } from './broadcast/listeners.js';
-import { streamStatus } from './broadcast/liquidsoap-control.js';
-import {
-  primaryLeg,
-  fallbackLeg,
-  probeLegReachable,
-  providerName,
-  activeModelLabel,
-} from './llm/provider.js';
-import { recentCalls } from './llm/log.js';
-import { getSetupStatus } from './setup/firstRun.js';
-import { djObject } from './llm/sdk.js';
-import { searchReady, searchWeb } from './skills/web-search.js';
-import * as system from './system.js';
-import { budgetStatus } from './broadcast/dj-budget.js';
-import * as analyzer from './music/analyzer.js';
-import { DJ_DOC_KNOWLEDGE } from './doctor-knowledge.js';
+import type { DoctorReport, DoctorSection, Finding, StationSettings } from './doctor/types.js';
+import { safe } from './doctor/util.js';
+import { rememberReport } from './doctor/cache.js';
+import { checkBroadcast, checkLlm, checkNavidrome, checkTts } from './doctor/checks-services.js';
+import { checkCapabilities, checkResources, checkTuning } from './doctor/checks-station.js';
+import { checkContent, checkSetup, checkStorage } from './doctor/checks-content.js';
 
-export type Status = 'ok' | 'warn' | 'fail' | 'skip';
-
-// Closed set of safe, already-implemented admin actions a finding may offer as a
-// one-click remediation. The web panel maps each id → the matching POST route.
-export type FixId =
-  | 'refresh-playlist'
-  | 'restart-mixer'
-  | 'generate-jingles'
-  | 'tag-library'
-  | 'subsonic-reset';
-
-export interface FixAction {
-  id: FixId;
-  label: string;
-}
-
-export interface Finding {
-  label: string;
-  status: Status;
-  detail?: string;
-  hint?: string;
-  fix?: FixAction;
-}
-
-export interface DoctorSection {
-  name: string;
-  findings: Finding[];
-}
-
-export interface DoctorReport {
-  t: string;
-  sections: DoctorSection[];
-  counts: { ok: number; warn: number; fail: number; skip: number };
-}
-
-export interface ReviewPriority {
-  title: string;
-  severity: 'low' | 'med' | 'high';
-  why: string;
-  suggestedFix: string;
-  // When set, the panel can render the matching one-click fix button right on
-  // this priority. The panel still cross-checks it against the report's own
-  // findings, so a stray/hallucinated id never surfaces a button.
-  fixId?: FixId | null;
-}
-
-export interface DoctorReview {
-  available: boolean;
-  reason?: string; // why the review couldn't run (LLM offline / not configured)
-  overall?: 'healthy' | 'attention' | 'critical';
-  summary?: string;
-  priorities?: ReviewPriority[];
-}
-
-// The performance-affecting slice of settings.get() the doctor inspects. All
-// optional — settings.get() is untyped and any field may be absent — so every
-// read is a safe `?.` access. Widened over time as new checks reference more.
-interface StationSettings {
-  llm?: {
-    pickerAgent?: boolean;
-    reasoning?: boolean;
-    agentTimeoutMs?: number;
-    provider?: string;
-    numCtx?: number;
-    maxOutputTokens?: number;
-    requestWebResolve?: boolean;
-    toolChoice?: string;
-    exemptRequests?: boolean;
-    pauseWhenEmpty?: boolean;
-    noRepeatWindow?: number;
-  };
-  tts?: { defaultEngine?: string; byKind?: Record<string, string | undefined> };
-  search?: { provider?: string };
-  audio?: { embeddings?: boolean; vocalActivity?: boolean; stemCache?: boolean };
-  transitions?: { pairDrain?: boolean; stemBlends?: boolean };
-  stream?: { opusEnabled?: boolean; flacEnabled?: boolean; aacEnabled?: boolean };
-  archive?: { enabled?: boolean };
-  crossfadeDuration?: number;
-  jingleRatio?: number;
-  maxTrackSeconds?: number;
-  loudness?: { targetLufs?: number };
-}
-
+// Re-exported so `import * as doctor from './doctor.js'` still reaches the
+// whole surface.
+export * from './doctor/types.js';
+export * from './doctor/cache.js';
+export { clearNavidromeCache, navidromeConnectivity } from './doctor/checks-services.js';
+export { reviewReport } from './doctor/review.js';
 // ---------------------------------------------------------------------------
 // runDoctor — assemble all sections.
 // ---------------------------------------------------------------------------
@@ -178,1068 +93,4 @@ export async function runDoctor(): Promise<DoctorReport> {
   return finalizeReport(sections);
 }
 
-// ---------------------------------------------------------------------------
-// Last-run cache — the panel hydrates from this on mount and the admin header
-// badges station health from it, so a report survives navigation and a restart
-// without forcing an immediate re-run. Deterministic data only (counts + DJ
-// Doc's verdict when a review has run). Best-effort disk mirror under STATE_DIR;
-// never throws into a check path.
-// ---------------------------------------------------------------------------
 
-const CACHE_FILE = `${STATE_DIR}/doctor-last.json`;
-
-interface DoctorCache {
-  report: DoctorReport | null;
-  review: DoctorReview | null;
-}
-
-let cache: DoctorCache = { report: null, review: null };
-let cacheLoaded = false;
-
-async function ensureCacheLoaded(): Promise<void> {
-  if (cacheLoaded) return;
-  cacheLoaded = true;
-  try {
-    const parsed = JSON.parse(await readFile(CACHE_FILE, 'utf8'));
-    if (parsed && typeof parsed === 'object') {
-      cache = { report: parsed.report ?? null, review: parsed.review ?? null };
-    }
-  } catch { /* no prior run persisted — fine */ }
-}
-
-function persistCache(): void {
-  // Fire-and-forget: the in-memory cache is the source of truth for this
-  // session; the disk mirror only bootstraps the next boot.
-  writeFile(CACHE_FILE, JSON.stringify(cache)).catch(() => {});
-}
-
-export function rememberReport(report: DoctorReport): void {
-  cacheLoaded = true; // a fresh run supersedes anything still on disk
-  // A new report describes a new moment — the old review no longer matches it.
-  cache = { report, review: null };
-  persistCache();
-}
-
-export function rememberReview(review: DoctorReview): void {
-  if (!cache.report) return; // a review with no report to attach to is meaningless
-  cache = { ...cache, review };
-  persistCache();
-}
-
-// Derive a headline verdict from the counts alone — used for the header badge
-// when no LLM review has run (runDoctor is LLM-free). A real review's `overall`
-// takes precedence when present.
-export function derivedOverall(counts: DoctorReport['counts']): 'healthy' | 'attention' | 'critical' {
-  if (counts.fail > 0) return 'critical';
-  if (counts.warn > 0) return 'attention';
-  return 'healthy';
-}
-
-export interface DoctorSummary {
-  t: string | null;
-  counts: DoctorReport['counts'] | null;
-  overall: 'healthy' | 'attention' | 'critical' | null;
-}
-
-// Compact health headline for the admin header badge — no section detail.
-export async function lastSummary(): Promise<DoctorSummary> {
-  await ensureCacheLoaded();
-  const r = cache.report;
-  if (!r) return { t: null, counts: null, overall: null };
-  const overall =
-    cache.review?.available && cache.review.overall ? cache.review.overall : derivedOverall(r.counts);
-  return { t: r.t, counts: r.counts, overall };
-}
-
-// Full last-run payload for the panel to hydrate from on mount.
-export async function lastRun(): Promise<DoctorCache> {
-  await ensureCacheLoaded();
-  return cache;
-}
-
-// Wrap a check so a thrown error becomes a single fail finding rather than
-// aborting runDoctor.
-async function safe(fn: () => Promise<Finding[]>): Promise<Finding[]> {
-  try {
-    return await fn();
-  } catch (err) {
-    return [{ label: 'check failed', status: 'fail', detail: err?.message || String(err) }];
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Sections
-// ---------------------------------------------------------------------------
-
-async function checkLlm(s: StationSettings | null): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  // Primary leg. probeLegReachable returns true for cloud providers (no cheap
-  // probe) and only false on a connection failure for local hosts.
-  try {
-    const leg = primaryLeg();
-    const ok = await probeLegReachable(leg);
-    out.push({
-      label: 'provider',
-      status: ok ? 'ok' : 'fail',
-      detail: `${providerName()} · ${activeModelLabel()}${ok ? ' · reachable' : ' · unreachable'}`,
-      hint: ok
-        ? undefined
-        : 'Without the LLM the DJ falls back to a stateless picker and skips spoken links. Check the provider, model and host in Settings → LLM.',
-    });
-  } catch (err) {
-    out.push({
-      label: 'provider',
-      status: 'fail',
-      detail: err?.message || 'not configured',
-      hint: 'Pick a provider + model in Settings → LLM.',
-    });
-  }
-
-  // Fallback leg (optional).
-  try {
-    const fb = fallbackLeg();
-    if (fb) {
-      const ok = await probeLegReachable(fb);
-      out.push({
-        label: 'fallback',
-        status: ok ? 'ok' : 'warn',
-        detail: ok ? 'configured · reachable' : 'configured · unreachable',
-      });
-    } else {
-      out.push({ label: 'fallback', status: 'skip', detail: 'none configured (optional)' });
-    }
-  } catch { /* fallback is best-effort */ }
-
-  // Recent error rate from the in-memory ring.
-  const recent = recentCalls.slice(0, 20);
-  if (recent.length) {
-    const fails = recent.filter((c) => c && c.ok === false).length;
-    const rate = Math.round((fails / recent.length) * 100);
-    out.push({
-      label: 'recent calls',
-      status: rate === 0 ? 'ok' : rate < 30 ? 'warn' : 'fail',
-      detail: `${fails}/${recent.length} failed (${rate}%)`,
-      hint:
-        rate >= 30
-          ? 'High failure rate. Confirm the model is loaded and the host is responsive (Debug → recent LLM calls has the errors).'
-          : undefined,
-    });
-  } else {
-    out.push({ label: 'recent calls', status: 'skip', detail: 'no calls yet' });
-  }
-
-  // Structured-output health — the silent failure mode behind "the model
-  // responds but features quietly break". djObject calls (DJ Doc's own AI review,
-  // the request matcher, the pool picker, the library tagger) need the model to
-  // emit JSON matching a strict shape; a weak model returns the wrong shape, the
-  // call fails Zod validation and the feature degrades or falls back unnoticed.
-  // We catch it deterministically here precisely because a model this broken ALSO
-  // breaks the AI review that would otherwise explain it to the operator.
-  const schemaFails = recentCalls.filter(isSchemaFailure);
-  if (schemaFails.length) {
-    const kinds = [...new Set(schemaFails.map((c) => c.kind).filter(Boolean))];
-    out.push({
-      label: 'structured output',
-      status: schemaFails.length >= 3 ? 'fail' : 'warn',
-      detail: `${schemaFails.length} schema-validation failure(s)${kinds.length ? ` · ${kinds.join(', ')}` : ''}`,
-      hint:
-        'The model is returning JSON that does not match the required shape, so these features fall back or go silent (DJ Doc’s own AI review is one of them). It’s the classic sign of a model that’s weak at schema-constrained output — usually a code-specialised or very small model. Switch Settings → LLM to a general instruction-tuned model (a ~12B+ local or a capable cloud model), and try turning reasoning OFF — “thinking” output can corrupt the JSON.',
-    });
-  }
-
-  // Model class — weigh the chosen model's *name* against how it's being used.
-  // Heuristic only (name-based), so it never fails, only warns: a code-specialised
-  // model is tuned for programming rather than DJ links / structured picks, and a
-  // small model paired with the agentic picker tends to time out into the pool.
-  const cls = classifyModel(activeModelLabel());
-  if (cls.code) {
-    out.push({
-      label: 'model class',
-      status: 'warn',
-      detail: `${activeModelLabel()} looks code-specialised`,
-      hint: 'Code models are tuned for programming, not natural-language DJ links or schema-constrained JSON — they tend to write stiff intros and fail structured picks (the request matcher, pool picker and this very report). Prefer a general instruction-tuned model in Settings → LLM.',
-    });
-  } else if (cls.sizeB !== null && cls.sizeB < 11 && s?.llm?.pickerAgent !== false) {
-    out.push({
-      label: 'model class',
-      status: 'warn',
-      detail: `~${cls.sizeB}B model with the agentic picker on`,
-      hint: 'The agentic picker wants a ~12B-class (or good cloud) model; smaller models often time out into the pool fallback or fail structured picks. Either pick a larger model, or turn the agentic picker OFF (Settings → LLM) to use the simpler, more forgiving pool picker.',
-    });
-  }
-
-  // Picker agent toggle — off is valid (stateless picker) but worth surfacing.
-  // DJ Doc weighs this against the model size + host resources in its review.
-  out.push({
-    label: 'picker agent',
-    status: s?.llm?.pickerAgent === false ? 'warn' : 'ok',
-    detail: s?.llm?.pickerAgent === false ? 'off — stateless pool picker' : 'on — session DJ agent (wants ~12B+ / good cloud model)',
-  });
-
-  // Chain-of-thought (reasoning) — on costs latency + tokens; only worth it on a
-  // capable model where link quality beats speed. Surfaced so DJ Doc can advise.
-  out.push({
-    label: 'chain-of-thought',
-    status: 'ok',
-    detail: s?.llm?.reasoning ? 'reasoning ON (thinking models; slower, pricier)' : 'reasoning OFF (faster, cheaper — good for small/local models)',
-  });
-
-  // Agent deadline — the wall-clock budget before the agentic picker falls back
-  // to the pool. Reasoning/cloud models routinely need 20–40s.
-  const deadlineMs = Number(s?.llm?.agentTimeoutMs);
-  if (Number.isFinite(deadlineMs) && deadlineMs > 0) {
-    out.push({
-      label: 'agent deadline',
-      status: 'ok',
-      detail: `${Math.round(deadlineMs / 1000)}s before falling back to the pool`,
-      hint: deadlineMs < 20000
-        ? 'Tight — reasoning-heavy or cloud models routinely need 20–40s, so the agent may keep falling back. Raise it if you run a slow model.'
-        : undefined,
-    });
-  }
-
-  return out;
-}
-
-async function checkNavidrome(): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  const p = await subsonic.ping();
-  out.push({
-    label: 'connectivity',
-    status: p.ok ? 'ok' : 'fail',
-    detail: p.ok ? `${config.navidrome.url} · authenticated` : p.reason || 'unreachable',
-    hint: p.ok
-      ? undefined
-      : 'The picker has no music source without Navidrome. Check the URL / username / password in setup, and that Navidrome is up.',
-  });
-
-  // Recent call error rate across all endpoints.
-  try {
-    const snap = subsonicLog.snapshot();
-    const calls = snap.endpoints.reduce((n: number, e) => n + e.calls, 0);
-    const errs = snap.endpoints.reduce((n: number, e) => n + e.errors, 0);
-    if (calls > 0) {
-      const rate = Math.round((errs / calls) * 100);
-      out.push({
-        label: 'call errors',
-        status: rate === 0 ? 'ok' : rate < 10 ? 'warn' : 'fail',
-        detail: `${errs}/${calls} calls errored (${rate}%)`,
-        fix: rate > 0 ? { id: 'subsonic-reset', label: 'Reset stats' } : undefined,
-      });
-    } else {
-      out.push({ label: 'call errors', status: 'skip', detail: 'no calls yet' });
-    }
-  } catch { /* tracker is best-effort */ }
-
-  // Mood-tag coverage — the picker leans on these tags to match the vibe.
-  try {
-    await library.load();
-    const st = library.stats();
-    out.push({
-      label: 'mood-tag coverage',
-      status: st.total > 0 ? 'ok' : 'warn',
-      detail:
-        st.total > 0
-          ? `${st.total} tracks tagged · ${st.distinctArtists} artists`
-          : 'no tracks tagged yet',
-      hint:
-        st.total > 0
-          ? undefined
-          : 'The picker matches tracks to the time-of-day / weather mood via these tags. Tag the library so it has something to work with.',
-      fix: st.total === 0 ? { id: 'tag-library', label: 'Tag library' } : undefined,
-    });
-  } catch (err) {
-    out.push({ label: 'mood-tag coverage', status: 'skip', detail: err?.message || 'library unavailable' });
-  }
-
-  // Embedding model perf advisory — a heavy LOCAL embedding model (bge-m3,
-  // *-large) is the quiet cause of slow re-embeds + Ollama RAM thrash on a
-  // CPU/NAS box. Deterministic + name-based (no probe), so it only ever warns.
-  try {
-    const adv = embeddings.embeddingPerfAdvisory();
-    const flag = adv.heavy && adv.local;
-    out.push({
-      label: 'embedding model',
-      status: flag ? 'warn' : 'ok',
-      detail: `${adv.provider}:${adv.model}${flag ? ' · heavy for a local host' : ''}`,
-      hint: flag
-        ? 'This is a large local embedding model — roughly 3–4× the size and 2–3× slower per track than the default nomic-embed-text, with bigger vectors (slower KNN, more RAM). On a CPU / NAS host it dominates re-embed time and can thrash Ollama when RAM is tight (it reloads the model between calls). Unless you specifically need its multilingual / long-context quality, switch Settings → Library tagger → Embedding to nomic-embed-text, then re-embed (Library → Maintenance → Re-embed all tracks).'
-        : undefined,
-    });
-  } catch { /* embedding cfg unavailable — skip silently */ }
-
-  return out;
-}
-
-// Cached live-config Navidrome connectivity for the always-on admin banner.
-// The banner polls this from every admin page every ~30s; a short cache keeps
-// that from becoming a steady drip of Subsonic `ping` calls (and shields a
-// flapping Navidrome). Shares subsonic.ping() — the same never-throwing check
-// checkNavidrome() uses — so the banner and the Doctor's connectivity finding
-// can never disagree.
-let navidromeCache: { at: number; result: { ok: boolean; reason?: string } } | null = null;
-const NAVIDROME_TTL_MS = 20_000;
-
-export async function navidromeConnectivity(): Promise<{
-  ok: boolean;
-  reason?: string;
-  url: string;
-}> {
-  const now = Date.now();
-  if (!navidromeCache || now - navidromeCache.at > NAVIDROME_TTL_MS) {
-    navidromeCache = { at: now, result: await subsonic.ping() };
-  }
-  return { ...navidromeCache.result, url: config.navidrome.url };
-}
-
-// Drop the cached ping so the banner/Doctor re-probe immediately — called when
-// the admin saves new Navidrome creds, where a stale "down" result would keep
-// the red banner up for the TTL even though the fix just landed.
-export function clearNavidromeCache() {
-  navidromeCache = null;
-}
-
-async function checkBroadcast(): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  // Icecast — is anything actually being served, and to whom.
-  try {
-    const st = getStreamStatus();
-    out.push({
-      label: 'Icecast stream',
-      status: st.online ? 'ok' : 'fail',
-      detail: st.online
-        ? `online · ${st.listeners?.current ?? 0} listening · ${st.bitrate ?? '?'}kbps`
-        : 'offline — nothing on /stream.mp3',
-      hint: st.online
-        ? undefined
-        : 'Liquidsoap may have dropped its Icecast connection. A mixer restart reconnects it.',
-      fix: st.online ? undefined : { id: 'restart-mixer', label: 'Restart mixer' },
-    });
-  } catch (err) {
-    out.push({ label: 'Icecast stream', status: 'skip', detail: err?.message || 'status unavailable' });
-  }
-
-  // Liquidsoap telnet — proves the mixer process is alive and reachable.
-  try {
-    const on = await streamStatus();
-    out.push({
-      label: 'mixer (Liquidsoap)',
-      status: on ? 'ok' : 'warn',
-      detail: on ? 'telnet reachable · stream on' : 'telnet reachable · stream off',
-      fix: on ? undefined : { id: 'restart-mixer', label: 'Restart mixer' },
-    });
-  } catch (err) {
-    out.push({
-      label: 'mixer (Liquidsoap)',
-      status: 'fail',
-      detail: `telnet unreachable: ${err?.message || 'no response'}`,
-      hint: 'The mixer process may be down or restarting. Check broadcast logs.',
-      fix: { id: 'restart-mixer', label: 'Restart mixer' },
-    });
-  }
-
-  return out;
-}
-
-async function checkTts(s: StationSettings | null): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  let avail: Record<string, unknown> = {};
-  try { avail = tts.availableEngines(); } catch { avail = {}; }
-
-  // Which engines the operator wants vs. which are actually available. A
-  // configured-but-unavailable engine silently falls back to Piper.
-  const wanted = new Set<string>();
-  const def = s?.tts?.defaultEngine;
-  if (def) wanted.add(def);
-  for (const v of Object.values(s?.tts?.byKind || {})) {
-    if (typeof v === 'string' && v) wanted.add(v);
-  }
-  if (wanted.size === 0) wanted.add('piper');
-
-  const unavailable = [...wanted].filter((e) => avail[e] === false);
-  out.push({
-    label: 'configured engines',
-    status: unavailable.length === 0 ? 'ok' : 'warn',
-    detail:
-      unavailable.length === 0
-        ? `${[...wanted].join(', ')} — available`
-        : `unavailable: ${unavailable.join(', ')} (will fall back to Piper)`,
-    hint:
-      unavailable.length === 0
-        ? undefined
-        : 'A configured voice engine is unavailable, so the DJ speaks in the Piper fallback voice. Enable the engine (e.g. the tts-heavy profile / cloud key) or pick an available one in Settings.',
-  });
-
-  // Is the current persona's voice silently routing through a fallback?
-  try {
-    const { spoken } = tts.describeRouting();
-    out.push({
-      label: 'active routing',
-      status: spoken.fellBack ? 'warn' : 'ok',
-      detail: spoken.fellBack
-        ? `requested ${spoken.requested ?? '?'} → using ${spoken.engine ?? '?'}`
-        : `${spoken.engine ?? 'piper'}`,
-    });
-  } catch { /* routing snapshot is best-effort */ }
-
-  return out;
-}
-
-async function checkCapabilities(s: StationSettings | null): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  // Web search — backs the DJ's artist-news segments. DuckDuckGo is keyless;
-  // Tavily needs a key. Report config readiness, then do a short live probe so
-  // "is it actually working?" is answered, not just "is it configured?".
-  const provider = s?.search?.provider || 'duckduckgo';
-  let ready = true;
-  try { ready = searchReady(); } catch { ready = true; }
-
-  if (!ready) {
-    out.push({
-      label: 'web search',
-      status: 'warn',
-      detail: `${provider} selected but no API key`,
-      hint: 'Artist-news segments can\'t fetch. Add a Tavily key (SEARCH_API_KEY) or switch to DuckDuckGo (keyless) in Settings → Search.',
-    });
-    return out;
-  }
-
-  // Live probe, raced against a 5s timeout so a hung provider can't stall the run.
-  let okLive = false;
-  let reason = '';
-  try {
-    await Promise.race([
-      searchWeb('SUB-WAVE radio diagnostic ping').then(() => { okLive = true; }),
-      new Promise((_, rej) => setTimeout(() => rej(new Error('timed out')), 5000)),
-    ]);
-  } catch (err) {
-    reason = err?.message || 'unreachable';
-  }
-  out.push({
-    label: 'web search',
-    status: okLive ? 'ok' : 'warn',
-    detail: okLive ? `${provider} · reachable` : `${provider} · ${reason}`,
-    hint: okLive
-      ? undefined
-      : 'The artist-news segment needs outbound web access. Check the controller can reach the internet / the search provider.',
-  });
-
-  return out;
-}
-
-async function checkResources(): Promise<Finding[]> {
-  const out: Finding[] = [];
-  try {
-    const sys = await system.summary();
-    const h = sys.host;
-    const gb = (b: number) => (b / (1024 ** 3)).toFixed(1);
-    const memPct = h.memTotal ? Math.round((h.memUsed / h.memTotal) * 100) : 0;
-
-    out.push({
-      label: 'CPU',
-      status: 'ok',
-      detail: `${h.cpus} cores · load ${h.loadavg.map((n) => n.toFixed(2)).join(' / ')}`,
-    });
-    out.push({
-      label: 'memory',
-      status: memPct > 90 ? 'warn' : 'ok',
-      detail: `${gb(h.memUsed)} / ${gb(h.memTotal)} GB used (${memPct}%)`,
-      hint: memPct > 90
-        ? 'Memory is tight — heavy TTS (Chatterbox / PocketTTS) and large local models may struggle. Prefer Piper/Kokoro and a smaller model.'
-        : undefined,
-    });
-    if (sys.dockerAvailable) {
-      const top = sys.containers[0];
-      out.push({
-        label: 'containers',
-        status: 'ok',
-        detail: top
-          ? `${sys.containers.length} running · busiest ${top.service} ${top.cpuPct}% CPU`
-          : `${sys.containers.length} running`,
-      });
-    }
-  } catch (err) {
-    out.push({ label: 'host resources', status: 'skip', detail: err?.message || 'unavailable' });
-  }
-  return out;
-}
-
-// Tuning — the cross-setting + telemetry-vs-config checks that the settings
-// panel itself can't show. Each finding weighs one performance knob against the
-// OTHER settings, the live LLM telemetry, or the host resources — the stuff an
-// operator can't eyeball. Findings are advisory (no one-click fix): the remedy
-// is a settings change, which DJ Doc's review then narrates the trade-off for.
-async function checkTuning(s: StationSettings | null): Promise<Finding[]> {
-  const out: Finding[] = [];
-  const llm: NonNullable<StationSettings['llm']> = s?.llm || {};
-
-  // --- Daily token budget: cliff detection + burn-rate projection ---
-  // Uses the live tally (broadcast/dj-budget), not just the configured number,
-  // so it can say "you'll hit the cap by mid-afternoon" instead of "a cap is set".
-  try {
-    const b = budgetStatus();
-    if (!b.enabled) {
-      out.push({ label: 'token budget', status: 'skip', detail: 'no daily cap (unlimited)' });
-    } else {
-      // A soft tier of 0 or 100 disables graceful degrade — the DJ jumps from
-      // full service straight to silent at the cap.
-      if (b.softPct <= 0 || b.softPct >= 100) {
-        out.push({
-          label: 'budget soft tier',
-          status: 'warn',
-          detail: `soft tier disabled (softPct ${b.softPct})`,
-          hint: 'With no soft tier the DJ jumps straight from full service to silent at the cap. Set Settings → LLM → budget soft % to ~80 so it first drops to the cheap pool picker and mutes optional segments, warning you before it hits the wall.',
-        });
-      }
-      // Project today's exhaustion from the burn rate so far (UTC day).
-      const now = Date.now();
-      const dayStart = Date.parse(new Date().toISOString().slice(0, 10) + 'T00:00:00Z');
-      const hrsElapsed = Math.max(0.05, (now - dayStart) / 3600000);
-      const rate = b.usedToday / hrsElapsed; // tokens/hour
-      const willExhaust = b.mode !== 'hard' && rate * 24 > b.cap && b.usedToday > 0;
-      const hitHour = willExhaust ? Math.min(24, Math.round(b.cap / rate)) : null;
-      const pct = b.cap > 0 ? Math.round((b.usedToday / b.cap) * 100) : 0;
-      out.push({
-        label: 'token budget',
-        status: b.mode === 'hard' ? 'fail' : b.mode === 'soft' || willExhaust ? 'warn' : 'ok',
-        detail:
-          b.mode === 'hard'
-            ? `cap reached (${fmtTokens(b.usedToday)}/${fmtTokens(b.cap)}) — DJ is coasting on the auto playlist`
-            : `${fmtTokens(b.usedToday)}/${fmtTokens(b.cap)} used today (${pct}%)${willExhaust ? ` · on track to hit the cap ~${hitHour}:00 UTC` : ''}`,
-        hint:
-          b.mode === 'hard'
-            ? 'No model calls until UTC midnight (listener requests still run if exempt). Raise Settings → LLM → daily token cap, or turn on pause-when-empty so idle hours don’t burn the budget.'
-            : willExhaust
-              ? 'At the current rate the DJ drops to the cheap picker (soft), then goes silent (hard) before the day ends. Raise the cap, enable pause-when-empty, or ease off reasoning / the agentic picker to slow the burn.'
-              : undefined,
-      });
-    }
-  } catch { /* budget module best-effort */ }
-
-  // --- Agent deadline vs MEASURED latency ---
-  // checkLlm only warns when the configured number looks low; here we compare it
-  // to what the model is actually doing. If p90 latency crowds the deadline, the
-  // agentic picker is silently timing out into the pool on most tracks.
-  if (llm.pickerAgent !== false) {
-    const lat = recentCalls
-      .filter((c) => c && c.ok !== false && Number.isFinite(c.ms) && c.ms > 0)
-      .map((c) => c.ms as number)
-      .sort((a, b) => a - b);
-    const deadlineMs = Number(llm.agentTimeoutMs);
-    if (lat.length >= 5 && Number.isFinite(deadlineMs) && deadlineMs > 0) {
-      const p90 = lat[Math.min(lat.length - 1, Math.floor(lat.length * 0.9))];
-      const ratio = p90 / deadlineMs;
-      out.push({
-        label: 'agent deadline vs latency',
-        status: ratio >= 1 ? 'fail' : ratio >= 0.8 ? 'warn' : 'ok',
-        detail: `p90 model latency ${(p90 / 1000).toFixed(1)}s vs ${Math.round(deadlineMs / 1000)}s deadline`,
-        hint:
-          ratio >= 0.8
-            ? 'Recent calls are running close to (or past) the agent deadline, so the agentic picker is probably timing out into the pool fallback on many tracks — you lose the session-aware picks you’re paying for. Raise Settings → LLM → agent deadline, or switch to a faster/smaller model (or turn reasoning OFF).'
-            : undefined,
-      });
-    }
-  }
-
-  // --- Context window too small for the agentic picker (Ollama only) ---
-  if (llm.provider === 'ollama' && llm.pickerAgent !== false) {
-    const nc = Number(llm.numCtx);
-    if (Number.isFinite(nc) && nc > 0 && nc < 8192) {
-      out.push({
-        label: 'context window',
-        status: 'warn',
-        detail: `numCtx ${nc} — tight for the agentic picker`,
-        hint: 'The agentic picker sends a system prompt + tool definitions + candidates; a small Ollama context window truncates them so the agent can’t call its “done” tool and falls back to the pool. Raise Settings → LLM → context window to ≥8192 (16384 is the default), or turn the agentic picker off.',
-      });
-    }
-  }
-
-  // --- Reasoning vs structured output (cross-setting, telemetry-backed) ---
-  // "Thinking" tokens are a common cause of malformed JSON. Tie the risk to the
-  // actual schema-failure count so it only fires when both are true.
-  if (llm.reasoning) {
-    const schemaFails = recentCalls.filter(isSchemaFailure).length;
-    if (schemaFails > 0) {
-      out.push({
-        label: 'reasoning vs structured output',
-        status: 'warn',
-        detail: `reasoning ON with ${schemaFails} recent schema failure(s)`,
-        hint: 'Chain-of-thought output is a common cause of malformed structured JSON (the picker, request matcher, tagger and this report all need strict JSON). Try turning Settings → LLM → reasoning OFF and re-checking — most local models emit cleaner JSON without it.',
-      });
-    }
-  }
-
-  // --- Max response size too small for the agentic picker ---
-  {
-    const mo = Number(llm.maxOutputTokens);
-    if (Number.isFinite(mo) && mo > 0 && mo < 1000 && llm.pickerAgent !== false) {
-      out.push({
-        label: 'max response size',
-        status: 'warn',
-        detail: `maxOutputTokens ${mo} — low for the agentic picker`,
-        hint: 'A small per-call output cap can truncate the agent’s tool calls mid-JSON, forcing a fallback. Leave Settings → LLM → max response size at 0 (strategy default) unless a small-context local model specifically needs it.',
-      });
-    }
-  }
-
-  // --- Listener-request web-resolve needs web search configured ---
-  if (llm.requestWebResolve) {
-    let ready = true;
-    try { ready = searchReady(); } catch { ready = true; }
-    if (!ready) {
-      out.push({
-        label: 'request web-resolve',
-        status: 'warn',
-        detail: 'on, but web search isn’t ready',
-        hint: 'Resolving described track requests ("that song from the film") needs a working web search. Configure it in Settings → Search, or turn Settings → LLM → request web-resolve off.',
-      });
-    }
-  }
-
-  // --- Audio-analysis features vs the analyzer flavour actually running ---
-  // audio.embeddings / vocalActivity silently no-op on the lean analyzer image
-  // (no CLAP / Demucs). Cross-check the setting against the live capability probe.
-  try {
-    const wantEmb = !!s?.audio?.embeddings;
-    const wantVocal = !!s?.audio?.vocalActivity;
-    if (wantEmb || wantVocal) {
-      try { await analyzer.refreshCapabilities(); } catch { /* best-effort probe */ }
-      // The upgrade path depends on which backend is running: 'sidecar' is the
-      // split stack's analyzer service; 'local' is the in-process venv — the
-      // AIO image (or a dev ANALYZE_PYTHON venv). Pointing an AIO operator at
-      // the analyzer image replaces their whole station with a bare analyzer
-      // micro-service (issue #966), so the hint has to name the right image.
-      const upgradeHint = analyzer.backendLabel() === 'local'
-        ? 'On the all-in-one image, switch the container to ghcr.io/perminder-klair/subwave-aio-heavy (NOT subwave-analyzer-heavy — that’s the bare analyzer service, not a station image); on a dev venv, install the heavy Python deps.'
-        : 'With docker compose, set ANALYZER_HEAVY=1 in .env and re-pull; without compose (Unraid, Portainer, plain docker) switch the analyzer container’s image to ghcr.io/perminder-klair/subwave-analyzer-heavy.';
-      if (wantEmb && analyzer.audioEmbeddingAvailable() === false) {
-        out.push({
-          label: 'audio embeddings',
-          status: 'warn',
-          detail: 'enabled, but the analyzer can’t produce them',
-          hint: `The “sounds-like” audio embeddings need the heavy analyzer (CLAP). You’re on the lean build, so this setting silently does nothing. ${upgradeHint} Or turn the setting off.`,
-        });
-      }
-      if (wantVocal && analyzer.vocalActivityAvailable() === false) {
-        out.push({
-          label: 'vocal activity',
-          status: 'warn',
-          detail: 'enabled, but the analyzer can’t produce it',
-          hint: `Vocal-range / talk-timing analysis needs the heavy analyzer (Demucs). The lean build can’t, so this setting no-ops. ${upgradeHint} Or turn the setting off.`,
-        });
-      }
-    }
-    // Stem cache / stem blends need the same Demucs stack.
-    const wantStems = !!s?.audio?.stemCache || !!s?.transitions?.stemBlends;
-    if (wantStems && analyzer.vocalActivityAvailable() === false) {
-      const upgradeHint = analyzer.backendLabel() === 'local'
-        ? 'On the all-in-one image, switch to ghcr.io/perminder-klair/subwave-aio-heavy; on a dev venv, install the heavy Python deps.'
-        : 'Set ANALYZER_HEAVY=1 in .env and re-pull, or switch the analyzer container to ghcr.io/perminder-klair/subwave-analyzer-heavy.';
-      out.push({
-        label: 'stem transitions',
-        status: 'warn',
-        detail: 'enabled, but the analyzer can’t separate stems',
-        hint: `The stem cache and stem-blend transitions need the heavy analyzer (Demucs). The lean build can’t separate, so these settings silently no-op. ${upgradeHint} Or turn them off.`,
-      });
-    }
-    // Stem blends ride the pair-drain scheduling — without it no seam is ever
-    // pair-annotated and a blend can never be inserted.
-    if (s?.transitions?.stemBlends && s?.transitions?.pairDrain === false) {
-      out.push({
-        label: 'stem transitions',
-        status: 'warn',
-        detail: 'stem blends are on but pair-drain scheduling is off',
-        hint: 'transitions.stemBlends needs transitions.pairDrain — a blend can only be inserted at a pair-aware drain. Re-enable pairDrain or turn stem blends off.',
-      });
-    }
-  } catch { /* analyzer unavailable — skip */ }
-
-  // --- Broadcast encoder load vs host CPU ---
-  // Every extra mount and the hourly archive is a continuous encoder — real CPU.
-  // Flag it only when several are on AND the host is already under pressure.
-  try {
-    const st: NonNullable<StationSettings['stream']> = s?.stream || {};
-    const extra: string[] = [];
-    if (st.opusEnabled) extra.push('opus');
-    if (st.flacEnabled) extra.push('flac');
-    if (st.aacEnabled) extra.push('aac');
-    if (s?.archive?.enabled) extra.push('hourly archive');
-    const encoderCount = 1 + extra.length; // /stream.mp3 always
-    const sys = await system.summary();
-    const cores = sys.host?.cpus || 0;
-    const load1 = Array.isArray(sys.host?.loadavg) ? sys.host.loadavg[0] : 0;
-    const perCore = cores > 0 ? load1 / cores : 0;
-    const pressured = (cores > 0 && cores <= 2) || perCore > 0.8;
-    const heavy = extra.length >= 2;
-    out.push({
-      label: 'broadcast encoders',
-      status: heavy && pressured ? 'warn' : 'ok',
-      detail: `${encoderCount} encoder${encoderCount === 1 ? '' : 's'} (mp3${extra.length ? ' + ' + extra.join(' + ') : ''}) · ${cores} cores, load ${load1.toFixed(2)}`,
-      hint:
-        heavy && pressured
-          ? 'Each extra stream mount and the hourly archive is a continuous encoder, and this host is already loaded for its core count. The hourly archive is the single biggest cost — disable mounts you don’t use (Settings → Broadcast) to free CPU for the DJ and TTS.'
-          : undefined,
-    });
-  } catch { /* resources unavailable — skip encoder check */ }
-
-  return out;
-}
-
-async function checkContent(): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  // auto.m3u — the fallback playlist Liquidsoap plays when the queue is empty.
-  const autoPath = `${STATE_DIR}/auto.m3u`;
-  out.push(await m3uFinding(autoPath, {
-    label: 'fallback playlist',
-    emptyHint: 'The autonomous fallback has nothing to play. Refresh it for the current mood.',
-    fix: { id: 'refresh-playlist', label: 'Refresh playlist' },
-  }));
-
-  // jingles.m3u — station idents.
-  out.push(await m3uFinding(`${STATE_DIR}/jingles.m3u`, {
-    label: 'jingles',
-    emptyHint: 'No station idents yet. Generate the defaults to get jingles between tracks.',
-    fix: { id: 'generate-jingles', label: 'Generate jingles' },
-  }));
-
-  return out;
-}
-
-// Shared helper: report an M3U as ok (N entries) / warn (empty or missing).
-async function m3uFinding(
-  path: string,
-  opts: { label: string; emptyHint: string; fix: FixAction },
-): Promise<Finding> {
-  try {
-    const body = await readFile(path, 'utf8');
-    const lines = body.split('\n').filter((l) => l.trim() && !l.startsWith('#'));
-    if (lines.length === 0) {
-      return { label: opts.label, status: 'warn', detail: 'empty', hint: opts.emptyHint, fix: opts.fix };
-    }
-    return { label: opts.label, status: 'ok', detail: `${lines.length} entries` };
-  } catch {
-    return { label: opts.label, status: 'warn', detail: 'missing', hint: opts.emptyHint, fix: opts.fix };
-  }
-}
-
-async function checkStorage(): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  const archive = await dirSize(`${STATE_DIR}/archive`);
-  const GB = 1024 * 1024 * 1024;
-  out.push({
-    label: 'hourly archive',
-    status: archive.bytes > 20 * GB ? 'warn' : 'ok',
-    detail: `${fmtBytes(archive.bytes)} across ${archive.files} files`,
-    hint:
-      archive.bytes > 20 * GB
-        ? 'The hourly archive is large. Prune old day folders under state/archive if disk is tight.'
-        : undefined,
-  });
-
-  const logs = await dirSize(`${STATE_DIR}/logs`);
-  out.push({
-    label: 'logs',
-    status: 'ok',
-    detail: `${fmtBytes(logs.bytes)} across ${logs.files} files`,
-  });
-
-  return out;
-}
-
-async function checkSetup(): Promise<Finding[]> {
-  const out: Finding[] = [];
-
-  try {
-    const st = await getSetupStatus();
-    out.push({
-      label: 'configuration',
-      status: st.needsSetup ? 'fail' : 'ok',
-      detail: st.needsSetup ? 'incomplete — Navidrome not configured' : `complete (${st.navidromeSource})`,
-      hint: st.needsSetup ? 'Finish the wizard at /onboarding (or run `subwave setup`).' : undefined,
-    });
-  } catch (err) {
-    out.push({ label: 'configuration', status: 'skip', detail: err?.message || 'unknown' });
-  }
-
-  // settings.get() throws if settings never loaded — surface that explicitly.
-  try {
-    settings.get();
-    out.push({ label: 'settings', status: 'ok', detail: 'loaded' });
-  } catch (err) {
-    out.push({ label: 'settings', status: 'fail', detail: err?.message || 'not loaded' });
-  }
-
-  // Gentle backup reminder — there's no signal to fail on, just good hygiene.
-  out.push({
-    label: 'backups',
-    status: 'skip',
-    detail: 'on demand',
-    hint: 'Export a backup from Admin → Backup so settings, personas, custom skills and library tags can be restored — especially before re-tagging or switching providers.',
-  });
-
-  return out;
-}
-
-// ---------------------------------------------------------------------------
-// reviewReport — hand the report to the LLM for a plain-English read.
-// ---------------------------------------------------------------------------
-
-const REVIEW_SCHEMA = z.object({
-  overall: z
-    .enum(['healthy', 'attention', 'critical'])
-    .describe('one-word verdict for the whole station right now'),
-  summary: z
-    .string()
-    .describe('2-4 sentences, plain English: how the station is doing and the single most important thing the operator should know'),
-  priorities: z
-    .array(
-      z.object({
-        title: z.string().describe('short imperative title of the thing to address'),
-        severity: z.enum(['low', 'med', 'high']),
-        why: z.string().describe('one sentence: why it matters for listeners or reliability'),
-        suggestedFix: z.string().describe('the concrete next step the operator should take'),
-        fixId: z
-          .enum(['refresh-playlist', 'restart-mixer', 'generate-jingles', 'tag-library', 'subsonic-reset'])
-          .nullable()
-          .optional()
-          .describe(
-            'When this priority is resolved by one of the one-click fixes listed in the prompt, you MUST set this to that exact fix id (so the operator gets a button). Leave it out only when no listed fix applies.',
-          ),
-      }),
-    )
-    .describe('0-5 items, highest severity first; empty array when everything is healthy'),
-});
-
-const REVIEW_SYSTEM = [
-  'You are DJ Doc — SUB/WAVE\'s resident station doctor. Picture a legendary West-Coast',
-  'hip-hop producer-engineer who has mixed a thousand records: a sharp ear, zero patience for a',
-  'muddy signal, and total confidence. You talk like a producer in the booth — a little swagger,',
-  'the odd studio metaphor ("the low end\'s clean", "that channel\'s clipping", "tighten the mix")',
-  '— but every call you make is technically correct. You review an automated health report for a',
-  'personal internet radio station (Navidrome library, an LLM DJ, Liquidsoap mixer, Icecast stream,',
-  'local/cloud TTS).',
-  'A knowledge base about how SUB/WAVE works is provided below — USE IT. Ground every recommendation',
-  'in it, and tailor model / picker / chain-of-thought / agent-deadline / budget / stream-encoder /',
-  'TTS-engine advice to the HOST RESOURCES and CURRENT SETTINGS shown in the report (e.g. don\'t tell',
-  'a small-CPU box to run Chatterbox, four stream encoders, or the agentic picker on a 9B model). The',
-  '"Tuning" section and the settings snapshot spell out the current knob values — use them to spot',
-  'settings that fight each other or the hardware, and say plainly which knob to change and to what.',
-  'Interpret the findings for a non-expert operator: what is fine, what needs attention, what to do',
-  'first. Be concrete and brief — keep the swagger to a light seasoning, no fluff, don\'t restate',
-  'every finding. Prioritise anything that takes the station off air or starves the DJ (stream',
-  'offline, Navidrome unreachable, LLM unreachable) above cosmetic warnings. Where it fits, nudge',
-  'the operator toward good hygiene like taking a backup. If everything is healthy, say so plainly',
-  'and return an empty priorities array.',
-].join(' ');
-
-export async function reviewReport(report: DoctorReport): Promise<DoctorReview> {
-  // Gate on reachability so we never hang the panel on a dead LLM host.
-  try {
-    const leg = primaryLeg();
-    const reachable = await probeLegReachable(leg);
-    if (!reachable) {
-      return { available: false, reason: `LLM host unreachable (${providerName()} · ${activeModelLabel()})` };
-    }
-  } catch (err) {
-    return { available: false, reason: err?.message || 'LLM not configured' };
-  }
-
-  try {
-    const review = await djObject({
-      system: REVIEW_SYSTEM,
-      prompt: [
-        DJ_DOC_KNOWLEDGE,
-        '\n---\n',
-        renderSettingsSnapshot(),
-        '',
-        'Here is the latest station health report.',
-        '',
-        renderReportText(report),
-        '',
-        renderAvailableFixes(report),
-        '',
-        'Review it for the operator. Lean on the knowledge base above. Cross-reference the "Tuning"',
-        'section and the current settings snapshot against the host resources, and tailor any model /',
-        'picker / chain-of-thought / agent-deadline / budget / encoder / TTS recommendations to what',
-        'this specific box can actually handle.',
-      ].join('\n'),
-      schema: REVIEW_SCHEMA,
-      temperature: 0.5,
-      kind: 'doctor:review',
-    });
-    const result: DoctorReview = { available: true, ...review };
-    rememberReview(result);
-    return result;
-  } catch (err) {
-    // A schema/shape mismatch here is itself a diagnosis: the review model can't
-    // do structured output. Say so plainly instead of dumping the Zod error, and
-    // point at the deterministic finding that survives a broken model.
-    const reason = isSchemaErrorMessage(err)
-      ? 'The review model returned output that doesn’t match the required shape — a sign the selected model is weak at structured output. See the LLM → “structured output” finding above; switching to a general instruction-tuned model fixes it.'
-      : err?.message || 'review failed';
-    return { available: false, reason };
-  }
-}
-
-// A failed LLM call whose error is a schema/shape mismatch (Zod) rather than a
-// host being unreachable — the fingerprint of a model that can't reliably produce
-// structured output. Matched on message text so it works across providers.
-function isSchemaFailure(c: { ok?: unknown; error?: unknown } | null | undefined): boolean {
-  return !!c && c.ok === false && isSchemaErrorMessage(c.error);
-}
-
-function isSchemaErrorMessage(err: unknown): boolean {
-  if (!err) return false;
-  const s = typeof err === 'string' ? err : (err as { message?: string }).message || JSON.stringify(err);
-  return /invalid_type|invalid_value|invalid_enum|unrecognized_keys|Invalid (option|input)|received undefined|No object generated|did not match (the )?schema|Type validation failed/i.test(s);
-}
-
-// Cheap name-based model classification — drives warnings only. `code` flags a
-// code-specialised model (poor at DJ links / structured output); `sizeB` is the
-// parameter count parsed from the tag (e.g. "gemma2:9b" → 9), null when absent.
-function classifyModel(label: string): { code: boolean; sizeB: number | null } {
-  const m = (label || '').toLowerCase();
-  const code = /coder|codestral|\bcode\b|[-_:]code/.test(m);
-  const sizeMatch = m.match(/(\d+(?:\.\d+)?)\s*b(?:[^a-z0-9]|$)/);
-  const sizeB = sizeMatch ? parseFloat(sizeMatch[1]) : null;
-  return { code, sizeB };
-}
-
-// Compact snapshot of the live performance-affecting settings, so the AI review
-// can reason about the actual knob values + trade-offs (not just the findings
-// that fired). Pairs with the "Tuning" section and the knowledge base.
-function renderSettingsSnapshot(): string {
-  let s: StationSettings | null = null;
-  try { s = settings.get(); } catch { return ''; }
-  const llm: NonNullable<StationSettings['llm']> = s?.llm || {};
-  const st: NonNullable<StationSettings['stream']> = s?.stream || {};
-  const mounts = ['mp3'];
-  if (st.opusEnabled) mounts.push('opus');
-  if (st.flacEnabled) mounts.push('flac');
-  if (st.aacEnabled) mounts.push('aac');
-
-  let budget = 'unlimited';
-  try {
-    const b = budgetStatus();
-    budget = b.enabled
-      ? `${fmtTokens(b.usedToday)}/${fmtTokens(b.cap)} today (soft ${b.softPct}%, mode ${b.mode})`
-      : 'unlimited';
-  } catch { /* budget best-effort */ }
-
-  let analyzerLabel = 'unknown';
-  try {
-    analyzerLabel = analyzer.backendLabel();
-    // Append the definitive CLAP capability so the AI review reasons from fact,
-    // not from "local probably means lean" (issue #966's false warning).
-    const clap = analyzer.audioEmbeddingAvailable();
-    analyzerLabel += clap === null ? ' (CLAP unknown)' : clap ? ' (heavy: CLAP yes)' : ' (lean: no CLAP)';
-  } catch { /* best-effort */ }
-
-  return [
-    '## Current settings (tune these against the findings + host resources)',
-    `- LLM: ${providerName()} · ${activeModelLabel()} · pickerAgent ${llm.pickerAgent === false ? 'OFF' : 'ON'} · reasoning ${llm.reasoning ? 'ON' : 'OFF'} · agentDeadline ${Math.round(Number(llm.agentTimeoutMs || 0) / 1000)}s · numCtx ${llm.numCtx ?? 'n/a'} · toolChoice ${llm.toolChoice || 'required'} · maxOutputTokens ${llm.maxOutputTokens || 'default'}`,
-    `- Budget: ${budget} · exemptRequests ${llm.exemptRequests === false ? 'OFF' : 'ON'} · pauseWhenEmpty ${llm.pauseWhenEmpty ? 'ON' : 'OFF'}`,
-    `- Picking: noRepeatWindow ${llm.noRepeatWindow ?? 'n/a'} · requestWebResolve ${llm.requestWebResolve ? 'ON' : 'OFF'}`,
-    `- Broadcast: crossfade ${s?.crossfadeDuration ?? '?'}s · jingle 1/${s?.jingleRatio ?? '?'} · maxTrack ${s?.maxTrackSeconds ? s.maxTrackSeconds + 's' : 'unlimited'} · loudness ${s?.loudness?.targetLufs ?? '?'} LUFS · mounts ${mounts.join('+')} · archive ${s?.archive?.enabled ? 'ON' : 'OFF'}`,
-    `- Voice: default TTS ${s?.tts?.defaultEngine || 'piper'}`,
-    `- Analysis: audio.embeddings ${s?.audio?.embeddings ? 'ON' : 'OFF'} · vocalActivity ${s?.audio?.vocalActivity ? 'ON' : 'OFF'} · analyzer ${analyzerLabel}`,
-    `- Search: ${s?.search?.provider || 'duckduckgo'}`,
-  ].join('\n');
-}
-
-// Tell the reviewer which one-click fixes are actually applicable right now —
-// the set of fix ids the findings surfaced this run — so it only attaches a
-// `fixId` to a priority when the corresponding button really exists.
-function renderAvailableFixes(report: DoctorReport): string {
-  const seen = new Map<FixId, string>();
-  for (const sec of report.sections) {
-    for (const f of sec.findings) {
-      if (f.fix && !seen.has(f.fix.id)) seen.set(f.fix.id, f.fix.label);
-    }
-  }
-  if (!seen.size) {
-    return 'One-click fixes available this run: none. Do NOT set fixId on any priority.';
-  }
-  const lines = [...seen.entries()].map(([id, label]) => `- ${id} — ${label}`);
-  return [
-    'One-click fixes available this run — for ANY priority that one of these resolves, you MUST set that',
-    'priority\'s `fixId` to the exact id shown here (the operator then gets a working button, not just',
-    'prose telling them to run it):',
-    ...lines,
-  ].join('\n');
-}
-
-// Compact, prompt-friendly rendering of the report — labels/status/detail/hint
-// only, no big blobs.
-function renderReportText(report: DoctorReport): string {
-  const lines: string[] = [];
-  lines.push(`Summary: ${report.counts.ok} ok, ${report.counts.warn} warn, ${report.counts.fail} fail, ${report.counts.skip} skip.`);
-  for (const sec of report.sections) {
-    lines.push(`\n## ${sec.name}`);
-    for (const f of sec.findings) {
-      let line = `- [${f.status.toUpperCase()}] ${f.label}`;
-      if (f.detail) line += `: ${f.detail}`;
-      if (f.hint) line += ` — hint: ${f.hint}`;
-      lines.push(line);
-    }
-  }
-  return lines.join('\n');
-}
-
-// ---------------------------------------------------------------------------
-// Small fs helpers
-// ---------------------------------------------------------------------------
-
-// Recursive directory size, capped so a huge archive can't make the check
-// expensive. Returns bytes + file count visited; missing dir → zeroes.
-async function dirSize(path: string, cap = 5000): Promise<{ bytes: number; files: number }> {
-  let bytes = 0;
-  let files = 0;
-  async function walk(dir: string): Promise<void> {
-    if (files >= cap) return;
-    let entries: Dirent[];
-    try {
-      entries = await readdir(dir, { withFileTypes: true });
-    } catch {
-      return;
-    }
-    for (const e of entries) {
-      if (files >= cap) return;
-      const full = `${dir}/${e.name}`;
-      if (e.isDirectory()) {
-        await walk(full);
-      } else {
-        try {
-          const s = await stat(full);
-          bytes += s.size;
-          files += 1;
-        } catch { /* file vanished mid-walk */ }
-      }
-    }
-  }
-  await walk(path);
-  return { bytes, files };
-}
-
-function fmtTokens(n: number): string {
-  if (!Number.isFinite(n)) return '?';
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-  if (n >= 1_000) return `${Math.round(n / 1_000)}k`;
-  return String(n);
-}
-
-function fmtBytes(n: number): string {
-  if (n < 1024) return `${n} B`;
-  const units = ['KB', 'MB', 'GB', 'TB'];
-  let v = n / 1024;
-  let i = 0;
-  while (v >= 1024 && i < units.length - 1) { v /= 1024; i++; }
-  return `${v.toFixed(1)} ${units[i]}`;
-}
